requirements.txt
opencv-python
Pillow
torch
transformers
matplotlib
scipy
streamlit
tensorflow